requirements.txt
django==3.0.8
celery==4.4.7
redis==3.5.3
flake8==3.8.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
black==19.10b0
psycopg2-binary==2.8.5